+++ src/main/java/Grammar/Expression/Variable.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 
 public record Variable(String name) implements Expression {
     @Override
-    public int eval(HashMap<String, Integer> bindings, Player p, Territory t) throws EvalError {
+    public double eval(HashMap<String, Integer> bindings, Player p, Territory t) throws EvalError {
         if (bindings.containsKey(name))
             return bindings.get(name);
         throw new EvalError("undefined variable: " + name);
